feat(wren-ui): refine adjust modal position
